paypal

+++ app/Http/Controllers/Api/BazarController.php
@@ -3,13 +3,14 @@
 namespace App\Http\Controllers\Api;
 
 use App\Models\Bazar;
+use Mockery\Undefined;
 use Illuminate\Http\Request;
 use App\Http\Requests\BazarRequest;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Repository\Bazar\BazarInterface;
 use App\Http\Resources\Unit\BazarResource;
 use App\Http\Traits\ResponseTrait as TraitResponseTrait;
-use Mockery\Undefined;
 
 class BazarController extends Controller
 {
@@ -19,6 +20,15 @@ class BazarController extends Controller
     public function __construct(BazarInterface $Repository)
     {
         $this->Repository = $Repository;
+
+        // $this->middleware(function ($request, $next) {
+        //     if (Auth::check() && Auth::user()->name !== 'مدخل البيانات') {
+        //         abort(403, 'Unauthorized');
+        //     }
+
+        //     return $next($request);
+        // });
+
     }
 
     public function index(Request $request)
@@ -30,12 +40,16 @@ class BazarController extends Controller
 
     public function store(BazarRequest $request)
     {
-        $data = $this->Repository->store($request->validated());
-        if ($data == true){
-            return $this->sendResponse($data, "تم تسجيل وحده جديده بنجاح", 200);
-        } else {
-            return $this->sendError($data, 'error', 404);
+        if (Auth::check() && Auth::user()->name == "مدخل البيانات ") {
 
+            $data = $this->Repository->store($request->validated());
+            if ($data == true) {
+                return $this->sendResponse($data, "تم تسجيل وحده جديده بنجاح", 200);
+            } else {
+                return $this->sendError($data, 'error', 404);
+            }
+        } else {
+            return $this->sendError('sorry', "you don't have permission to access this", 404);
         }
     }
 
@@ -49,11 +63,15 @@ class BazarController extends Controller
 
     public function storeUp(Request $request)
     {
-        $data = $this->Repository->edit($request);
-        if ($data === true){
-            return $this->sendResponse($data, "تم التعديل ", 200);
+        if (Auth::check() && Auth::user()->name == "مدخل البيانات ") {
+            $data = $this->Repository->edit($request);
+            if ($data === true) {
+                return $this->sendResponse($data, "تم التعديل ", 200);
+            } else {
+                return $this->sendError($data, 'error', 404);
+            }
         } else {
-            return $this->sendError($data, 'error', 404);
+            return $this->sendError('sorry', "you don't have permission to access this", 404);
         }
     }
 
@@ -61,15 +79,19 @@ class BazarController extends Controller
 
     public function destroy($id)
     {
-        return $this->sendResponse($this->Repository->delete($id), " تم حذف الوحده ", 200);
+        if (Auth::check() && Auth::user()->name == "مدخل البيانات ") {
+            return $this->sendResponse($this->Repository->delete($id), " تم حذف الوحده ", 200);
+        } else {
+            return $this->sendError('sorry', "you don't have permission to access this", 404);
+        }
     }
 
 
 
-    
+
     public function numbers()
     {
-        
+
         $numbers = Bazar::orderBy('number', 'asc')->get();
         $unique_data = $numbers->unique('number')->pluck('number')
             ->values()->all();
@@ -79,11 +101,11 @@ class BazarController extends Controller
             'data' => $unique_data
         ], 200);
     }
-    
-   
+
+
     public function revenue()
     {
-        
+
         $revenue = Bazar::orderBy('revenue', 'asc')->get();
         $unique_data = $revenue->unique('revenue')->pluck('revenue')
             ->values()->all();
@@ -127,6 +149,4 @@ class BazarController extends Controller
             'data' => $unique_data,
         ], 200);
     }
-
-
 }
